+++ bioconvert/abi2qual.py
@@ -10,20 +10,19 @@
 #  website: https://github.com/biokit/bioconvert
 #  documentation: http://bioconvert.readthedocs.io
 ##############################################################################
-"""Convert :term:`ABI` format to :term:`FASTQ` format"""
+"""Convert :term:`ABI` format to :term:`QUAL` format"""
 from bioconvert import ConvBase
 from bioconvert import requires
 
 
-__all__ = ["ABI2FASTQ"]
+__all__ = ["ABI2QUAL"]
 
 
-class ABI2FASTQ(ConvBase):
-    """Convert :term:`ABI` file to :term:`FASTQ` file
+class ABI2QUAL(ConvBase):
+    """Convert :term:`ABI` file to :term:`QUAL` file
 
-    :term:`ABI` files are created by ABI sequencing machine and includes
-    PHRED quality scores for base calls. This allows the creation
-    of :term:`FastQ` files.
+    :term:`ABI` files are created by ABI sequencing machine and includes PHREAD quality
+    scores for base calls. This allows the creation of :term:`QUAL` files. 
 
     """
 
@@ -33,13 +32,23 @@ class ABI2FASTQ(ConvBase):
         """.. rubric:: constructor
 
         :param str infile: input ABI file
-        :param str outfile: output FASTQ filename
+        :param str outfile: output QUAL filename
 
         """
-        super(ABI2FASTQ, self).__init__(infile, outfile, *args, **kargs)
+        super(ABI2QUAL, self).__init__(infile, outfile, *args, **kargs)
 
     @requires(python_library="biopython")
     def _method_biopython(self, *args, **kwargs):
         from Bio import SeqIO
         records = SeqIO.parse(self.infile, "abi")
-        SeqIO.write(records, self.outfile, "fastq")
+        # output using SeqIO.write(records, self.outfile, "qual") is not
+        # standard so we write our own conversion here below
+        with open(self.outfile, "w") as fout:
+            for rec in records:
+                header = rec.name
+                qual = rec.letter_annotations['phred_quality']
+                qual = "".join([str(x) for x in qual])
+                fout.write(">{}\n".format(header))
+                fout.write("{}\n".format(qual))
+
+