+++ bioconvert/sra2fastq.py
@@ -38,16 +38,23 @@ from bioconvert.core.decorators import requires
 class SRA2FASTQ(ConvBase):
     """Converts Sra 2 Fastq(.gz) file
 
+    ::
+
+        bioconvert sra2fastq ERR043367
+
+    This may take some times since the files are downloaded from SRA website.
+
     """
     _default_method = "sratoolkit"
 
     # If test: will take only the first 10 reads from the sra file
     def __init__(self, infile, outfile, test=False):
         """.. rubric:: constructor
+    
         :param str infile:
         :param str outfile:
-        library used::
-            sra-toolkit
+
+        library used: sra-toolkit
         """
         super().__init__(infile, outfile)
         self.test = test